Buoyancy term must be vertically cell-averaged.
According to Jean-Michel. This is because the T, S, p are defined as
averages over the cell volume so they are defined at the center of the
cells. Thus the density \rho(T,S,p) ends up being defined at the center
of the cells as well. However, the velocities are defined on the faces,
so to get the density at the interface/boundary/face between two cells
you take a vertical cell-average, i.e. avg^z from operators.jl.