Change default output array type for `JLD2OutputWriter` to `Array{Float64}` (#2890)
* Change default output array type for `JLD2OutputWriter` to Array{Float64}

While `Array{Float32}` is neat because user output is smaller, it's a bad default because it often leads to confusing results, like the fact that outputing `Field(Integral(c))` is different (possibly by `eps(Float32)`) than computing `Integral(c_output)` in post processing.

Thus @xkykai and I propose to make the default `Array{Float64}`. We could regard outputting with `Array{Float32}` as an "optimization" which is premature as a default.

* fix doctests

* Update netcdf_output_writer.jl

* fix more doctests

* fix tests

* Update test_netcdf_output_writer.jl

* Update test_netcdf_output_writer.jl

* fix tests

* Update Project.toml

* Update Project.toml

* Update Project.toml

---------

Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>