Add `SeawaterDensity` to `Oceananigans.Models` (#3329)
* First steps

* Simpler because `S` and `T` tracers are already required

What to do with `LinearEquationOfState` which does not have a reference density?

* For now restrict to `BoussinesqEquationOfState`

* Update density_model.jl

* Working, need tests

* Add tests

* Include the test script

Need to run the tests and look at some of my examples with a clean julia session before opening the PR

* Tests pass locally from cold start

* Only need export `SeawaterDensity`

* Extend to `SewaterBuoyancy` with constant `S` or `T`

Still needs to be a `BoussinesqEquationOfState`.

* Better docstrings

* Doctest and GPU fixes

Still not sure if `GPU()` tests will be fixed but I think array type should be correct. I ran `doctest` locally on the `Oceananigans.Models` module and there was no error

* Update src/Models/Models.jl

Co-authored-by: Gregory L. Wagner <gregory.leclaire.wagner@gmail.com>

* File rename to `seawater_density.jl`

* Simplify testing functions

* Changes after @glwagner comments

* Export in main + `KernelFunctionOperation` check to `AbstractOperation`

---------

Co-authored-by: Gregory L. Wagner <gregory.leclaire.wagner@gmail.com>
Co-authored-by: Gregory L. Wagner <wagner.greg@gmail.com>